remove Adobe Kuler integration
In prep for moving servers. It appears the integration has not worked
for at least a year, given that the API now redirects to HTTPS and
the server did not have recent-enough root certs to validate it.

Beyond that, Adobe Kuler has been shutdown (a placeholder response
appears to be hardcoded for graceful degredation), after a deprecation
phase in which it rebranded as Adobe Color.